+++ src/components/ToneGameScreen.tsx
@@ -25,6 +25,7 @@ import {
   playButtonTap,
   playPadTone,
 } from '../lib/sound'
+import { playCorrectHaptic, playIncorrectHaptic } from '../lib/haptics'
 import { SetSummary } from './SetSummary'
 import { GameHeader } from './GameHeader'
 import { ResultBadge } from './ResultBadge'
@@ -140,6 +141,10 @@ export function ToneGameScreen({
     if (loadSettings().soundEnabled) {
       if (correct) playCorrectSound()
       else playIncorrectSound()
+    }
+    if (loadSettings().hapticsEnabled) {
+      if (correct) playCorrectHaptic()
+      else playIncorrectHaptic()
     }
     setCurrentResult({ question: currentQuestion, tapped: value, correct })
     setPhase('result')